Moved Models into lib/models
