+++ lib/models/hotel.dart
