+++ lib/models/tour.dart
